backend/requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
